Implemented music and ambience system in FMOD
I've moved the music out of Unity Audio and into FMOD. It mostly works, except that its not handling scene changes at the end of the game and keeps loops playing. I've also exposed SFX, Music and Ambience volume separately in FmodPlayer, so they can be hooked up to sliders.